Send snapshotTransactionsEnabled attribute only if set explicitly by user
In Exasol 7.1.0, Snapshot Mode will be enabled by default, but that
default can also be changed using

  ALTER SYSTEM SET SNAPSHOT_MODE = 'OFF'|'SYSTEM TABLES';

Therefore the "default" for the WebSockets attribute
snapshotTransactionsEnabled can now be True (for 'SYSTEM TABLES') or
False (for 'OFF'), depending on the ALTER SYSTEM value.

Change pyexasol so that the attribute is only sent if the user
explicitly set it using the snapshot_transactions option.

Test case:

  # Setup
  # Connect to Exasol 7.1.0, with system default SYSTEM TABLES
  import pyexasol
  C1 = pyexasol.connect(dsn='localhost:45623', user='sys', password='exasol')
  C2 = pyexasol.connect(dsn='localhost:45623', user='sys', password='exasol', snapshot_transactions=False)

  # Existing behaviour: [('OFF',)]
  # Desired behaviour:  [('SYSTEM TABLES',)]
  print(C1.execute("select SESSION_VALUE from EXA_PARAMETERS where PARAMETER_NAME = 'SNAPSHOT_MODE'").fetchall())

  # Existing behaviour: [('OFF',)]
  # Desired behaviour:  [('OFF',)]
  print(C2.execute("select SESSION_VALUE from EXA_PARAMETERS where PARAMETER_NAME = 'SNAPSHOT_MODE'").fetchall())